requirements.txt
fastapi
uvicorn[standard]
openai
sentence-transformers
qdrant-client
pydantic
email-validator>=2.0
python-dotenv
